+++ src/backend/src/services/recruitment.services.ts
@@ -3,7 +3,7 @@ import { isAdmin, User } from 'shared';
 import prisma from '../prisma/prisma.js';
 import { AccessDeniedAdminOnlyException, DeletedException, NotFoundException } from '../utils/errors.utils.js';
 import { userHasPermission } from '../utils/users.utils.js';
-import { faqTransformer } from '../transformers/faq.transformer.js';
+import { faqTransformer, guestDefinitionTransformer } from '../transformers/recruitment-transformer.js';
 import { getFaqQueryArgs } from '../prisma-query-args/faq.query-args.js';
 
 export default class RecruitmentServices {
@@ -259,6 +259,6 @@ export default class RecruitmentServices {
       where: { organizationId: organization.organizationId, dateDeleted: null }
     });
 
-    return allGuestDefintions;
+    return allGuestDefintions.map(guestDefinitionTransformer);
   }
 }
