+++ src/backend/src/transformers/recruitment-transformer.ts
@@ -1,0 +1,23 @@
+import { Prisma } from '@prisma/client';
+import { FrequentlyAskedQuestion, GuestDefinition } from 'shared';
+import { FaqQueryArgs } from '../prisma-query-args/faq.query-args.js';
+import { userTransformer } from './user.transformer.js';
+
+export const faqTransformer = (faq: Prisma.FrequentlyAskedQuestionGetPayload<FaqQueryArgs>): FrequentlyAskedQuestion => ({
+  faqId: faq.faqId,
+  question: faq.question,
+  answer: faq.answer,
+  userCreated: userTransformer(faq.userCreated),
+  dateCreated: faq.dateCreated,
+  dateDeleted: faq.dateDeleted ?? undefined
+});
+
+export const guestDefinitionTransformer = (guestDefinition: Prisma.Guest_DefinitionGetPayload<{}>): GuestDefinition => ({
+  definitionId: guestDefinition.definitionId,
+  term: guestDefinition.term,
+  description: guestDefinition.description,
+  order: guestDefinition.order,
+  buttonText: guestDefinition.buttonText ?? undefined,
+  buttonLink: guestDefinition.buttonLink ?? undefined,
+  icon: guestDefinition.icon ?? undefined
+});
